+++ posix-configs/SITL/init/ekf2/hippocampus
@@ -13,7 +13,7 @@ dataman start
 
 simulator start -s
 
-pwm_out_sim mode_pwm
+pwm_out_sim start
 mixer load /dev/pwm_output0 ROMFS/px4fmu_test/mixers/uuv_quad_x.mix
 
 gyrosim start
