+++ posix-configs/SITL/init/ekf2/iris_rplidar
@@ -53,7 +53,7 @@ accelsim start
 barosim start
 adcsim start
 gpssim start
-pwm_out_sim mode_pwm
+pwm_out_sim start
 sensors start
 commander start
 land_detector start multicopter
