+++ posix-configs/SITL/init/ekf2/rover
@@ -57,7 +57,7 @@ barosim start
 adcsim start
 gpssim start
 measairspeedsim start
-pwm_out_sim mode_pwm
+pwm_out_sim start
 sleep 1
 sensors start
 commander start
