+++ posix-configs/SITL/init/ekf2/standard_vtol
@@ -68,7 +68,7 @@ barosim start
 adcsim start
 gpssim start
 measairspeedsim start
-pwm_out_sim mode_pwm
+pwm_out_sim start start
 sensors start
 commander start
 land_detector start vtol
