+++ posix-configs/SITL/init/ekf2/tailsitter
@@ -49,7 +49,7 @@ barosim start
 adcsim start
 gpssim start
 measairspeedsim start
-pwm_out_sim mode_pwm
+pwm_out_sim start
 sensors start
 commander start
 land_detector start vtol
