+++ posix-configs/SITL/init/lpe/iris_opt_flow
@@ -71,7 +71,7 @@ accelsim start
 barosim start
 adcsim start
 gpssim start
-pwm_out_sim mode_pwm
+pwm_out_sim start
 sensors start
 commander start
 land_detector start multicopter
