+++ posix-configs/SITL/init/lpe/iris_rplidar
@@ -52,7 +52,7 @@ accelsim start
 barosim start
 adcsim start
 gpssim start
-pwm_out_sim mode_pwm
+pwm_out_sim start
 sleep 1
 sensors start
 commander start
