+++ posix-configs/SITL/init/lpe/standard_vtol
@@ -63,7 +63,7 @@ barosim start
 adcsim start
 gpssim start
 measairspeedsim start
-pwm_out_sim mode_pwm
+pwm_out_sim start
 sensors start
 commander start
 land_detector start vtol
