+++ posix-configs/SITL/init/lpe/typhoon_h480
@@ -50,7 +50,7 @@ accelsim start
 barosim start
 adcsim start
 gpssim start
-pwm_out_sim mode_pwm16
+pwm_out_sim start
 sensors start
 commander start
 land_detector start multicopter
