+++ posix-configs/SITL/init/test/cmd_template.in
@@ -13,7 +13,7 @@ barosim start
 adcsim start
 gpssim start
 measairspeedsim start
-pwm_out_sim mode_pwm
+pwm_out_sim start
 
 ver all
