+++ posix-configs/eagle/hil/px4.config
@@ -47,7 +47,7 @@ mc_pos_control start
 mc_att_control start
 land_detector start multicopter
 sleep 1
-pwm_out_sim mode_pwm
+pwm_out_sim start
 mixer load /dev/pwm_output0 /startup/quad_x.main.mix
 list_devices
 list_files
