+++ posix-configs/rpi/px4_hil.config
@@ -18,7 +18,7 @@ mc_pos_control start
 mc_att_control start
 mavlink start -x -u 14577 -r 1000000
 navio_sysfs_rc_in start
-pwm_out_sim mode_pwm16
+pwm_out_sim start
 mixer load /dev/pwm_output0 ROMFS/px4fmu_common/mixers/quad_x.main.mix
 logger start -t -b 200
 mavlink boot_complete
